+++ r-package/R/download_metadata.R
@@ -17,23 +17,11 @@ download_metadata <- function(){ # nocov start
 
   } else {
 
-  # Download metadata
-
-    #supress warnings
-    oldw <- getOption("warn")
-    options(warn = -1)
-
     # test server connection
     metadata_link <- 'https://www.ipea.gov.br/geobr/metadata/metadata_gpkg.csv'
-    con <- url(metadata_link)
-    t <- suppressWarnings({ try( open.connection(con, open="rt", timeout=2), silent=T)[1] })
-    if ("try-error" %in% class(t)) {stop('Internet connection problem. If this is not a connection problem in your network, please try geobr again in a few minutes.')}
-    suppressWarnings({ try(close.connection(con), silent=TRUE) })
+    is_online(metadata_link)
 
-    # return with warnings
-    options(warn = oldw)
-
-    # download it and save to metadata
+    # download metadata to temp file
     httr::GET(url= metadata_link, httr::write_disk(tempf, overwrite = T))
 
     }