+++ r-package/NEWS.md
@@ -8,6 +8,7 @@
 * From v1.6 onwards, `geobr` stores downloaded in temporary cache in `tempdir()`. Closes #225.
 * New function `read_comparable_areas` to read historically comparable municipalities, aka Areas minimas comparaveis (AMCs). Closes issue #17
 * New data set of macro regions og health, which can be accessed using the new `macro` parameter added to `read_health_region()` function. Closes issue #219.
+* New internal support function `is_online()`to check internet connection with Ipea server. Closes #229
 * updated data sets for 2020:
   * Country
   * Regions